moved topic buttons to new file
