+++ src/shared/Routes.constants.tsx
@@ -10,6 +10,8 @@ export enum Routes {
   AdminCodeWithProblem = "/admin/code/:problemId?",
   AdminCode = "/admin/code",
   Code = "/code",
+  Learn = "/learn",
+  LearnWithLesson = "/learn:lessonId?",
   Landing = "/",
   PrivacyNotice = "/privacy_notice",
   PrivacyNoticeDocument = "/privacy_notice_document",
